removed the main.cpp in the top level folder.
