+++ Code/Emulator_debugging/main.cpp
@@ -1,29 +1,57 @@
 #include <iostream>
 #include "time.h"
-#include "emu.hpp"
-#include "InputHandler.hpp"
+#include <thread>
+#include <future>
+
+#include "Emulator_debugging/include/emu.hpp"
+#include "InputHandler/include/InputHandler.hpp"
+
+
+void getKey(std::promise<int> &prms) {
+    int key = getch();
+    prms.set_value(key);
+}
 
 int main(/*int argc, char ** argv*/){
-    clock_t current = clock();
+
     Emu emu;
     UserInput userInput;
     inputWindowInit();
 
     int input;
+    printw("START...");
 
-    while (1)
+    while (true)
     {
-        input = getch();
+        clock_t current = clock();
+        std::promise<int> prms;
+        auto ftr = prms.get_future();
+
+        std::thread thrIn(&getKey, std::ref(prms));
+
+        input = ftr.get();
+        thrIn.join();
+
         userInput.Sensing(input);
+
+        userInput.PrintSensorValues();
 
         int acc = userInput.getAccPedalPos();
 
+        std::cout << acc << std::endl;
+
         emu.currentTime = current;
-        
+        std::cout << current << std::endl;
+        std::cout << emu.currentTime << std::endl;
+
         emu.setAccPedalPos(acc);
+
         emu.updateRpm(emu.getAccPedalPos());
-        // emu.outputRpm(emu.getRpm());
 
         emu.lastTick = current;
+        for(int i =0; i<100000; i++) {
+
+        }
     }
+    return 0;
 }
